// generated/src/app/api-models/ar-cash-batch-control.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ArCashBatchControl {
  seqCashBatchId: number;
  itemCount: number;
  batchTotal: number;
  batchStatus: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}